connection now returns port num as short
port function returned an int.  it now returns a short.

+++ src/connection.h
@@ -32,7 +32,7 @@ public:
 	/**
 	 * Return the server port that this connection is on
 	 */
-	virtual int port() const = 0;
+	virtual short port() const = 0;
 
 	/**
 	 * Read a line of input from this connection.
